% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob-odds.R
\name{p_to_o}
\alias{p_to_o}
\title{Convert probability to odds.}
\usage{
p_to_o(probability, print = FALSE)
}
\arguments{
\item{probability}{A single numeric decimal probability (i.e., a number
between 0 and 1).}

\item{print}{A logical indicating whether to print a sentence description of
the calculated odds for the given probaiblity.}
}
\value{
The output will be decimal odds. If \code{print = TRUE}, the decimal
  odds will be returned invisibly.
}
\description{
\code{p_to_o()} converts decimal probability to odds.
}
\details{
If \code{print = TRUE}, an explanatory sentence will be printed in the
form "If the probability of an event occuring is <probability>, the odds
are <decimal odds> or <ratio:odds>."
}
\examples{
p_to_o(0.5)
p_to_o(0.1, print = TRUE)
calculated_odds <- p_to_o(0.9, print = TRUE)
calculated_odds
}
\seealso{
\code{\link{o_to_p}} for converting odds to probability.
}
