% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resids.R
\name{resid_fitted_plot}
\alias{resid_fitted_plot}
\title{Plot residuals vs. fitted values.}
\usage{
resid_fitted_plot(model, resid_type = "std", outlier_2sd_lines = TRUE,
  title = "Residuals vs. Fitted Values", subtitle = TRUE,
  xlab = "Fitted Values", ylab = "Residuals", ...)
}
\arguments{
\item{model}{An \code{lm} or \code{aov} model object.}

\item{resid_type}{A string indicating which type of residuals to use. The
default \code{"std"} uses standardized residuals, \code{"unstd"} uses
unstandardized residuals, and \code{"both"} produces both plots stacked
with \code{"unstd"} on top and \code{"std"} on the bottom.}

\item{outlier_2sd_lines}{A logical indicating whether to draw horizontal red
lines on the standardized residual plot at a y-value of +/- 2.}

\item{title}{A string indicating what to title the plot.}

\item{subtitle}{A logical indicating whether to include the default subtitle.}

\item{xlab}{A string indicating what to label the x-axis.}

\item{ylab}{A string indicating what to label the y-axis.}

\item{...}{Additional arguments passed to
\code{ggplot2::\link[ggplot2]{theme}()}.}
}
\value{
If a single plot type (\code{"std"} or \code{"unstd"}) is specified,
  a ggplot2 plot object is returned. If \code{"both"} plot types are
  requested, a call to \code{gridExtra::\link[gridExtra]{grid.arrange}} draws
  on the current device and a gtable is returned.
}
\description{
Plot model residuals vs. fitted values with ggplot2 to assess a model fit.
}
\examples{
mtcars_lm <- lm(mpg ~ wt, mtcars)
resid_fitted_plot(mtcars_lm)
resid_fitted_plot(mtcars_lm, resid_type = "unstd")
resid_fitted_plot(mtcars_lm, resid_type = "both")

}
\seealso{
\code{\link{resid_hist_qq}} for residual histograms and Q-Q plots.
}
