% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resids.R
\name{resid_hist_qq}
\alias{resid_hist_qq}
\title{Plot residual histograms and Q-Q plots.}
\usage{
resid_hist_qq(model, plot_type = "both", resid_type = "std", bins = 9,
  qqline = TRUE, htitle = "Histogram", qtitle = "Normal Q-Q Plot",
  subtitle = TRUE, hxlab = "Residuals", qxlab = "Theoretical Quantiles",
  hylab = "Count", qylab = "Sample Quantiles", ...)
}
\arguments{
\item{model}{An \code{lm} or \code{aov} model object.}

\item{plot_type}{A string indicating which type of plot to produce. The
default \code{"both"} produces both histograms and Q-Q plots side by side,
\code{"hist"} produces only histograms, and \code{"qq"} produces only Q-Q
plots.}

\item{resid_type}{A string indicating which type of residuals to use. The
default \code{"std"} uses standardized residuals, \code{"unstd"} uses
unstandardized residuals, and \code{"both"} produces both plots stacked
with \code{"unstd"} on top and \code{"std"} on the bottom.}

\item{bins}{How many bins to use for the histograms.}

\item{qqline}{A logical indicating whether to include a qqline on the Q-Q
plots.}

\item{htitle}{A string indicating what to title the histograms.}

\item{qtitle}{A string indicating what to title the Q-Q plots.}

\item{hxlab}{A string indicating what to title the histogram x-axes.}

\item{qxlab}{A string indicating what to title the Q-Q plot x-axes.}

\item{hylab}{A string indicating what to title the histogram y-axes.}

\item{qylab}{A string indicating what to title the Q-Q plot y-axes.}

\item{...}{Additional arguments passed to
\code{ggplot2::\link[ggplot2]{theme}()}.}

\item{subittle}{A logical indicating whether to include the default
subtitles.}
}
\value{
If single \code{plot_} and \code{resid_} \code{type}s are specified
  (i.e., a single histogram or Q-Q plot is requested), a ggplot2 plot object
  is returned. If more than one plot is requested, a call to
  \code{gridExtra::\link[gridExtra]{grid.arrange}} draws on the current
  device and a gtable is returned.
}
\description{
Plot histograms and Q-Q plots of model residuals to assess residual normality.
}
\examples{
mtcars_lm <- lm(mpg ~ wt, mtcars)
resid_hist_qq(mtcars_lm)
resid_hist_qq(mtcars_lm, plot_type = "qq", resid_type = "unstd")
resid_hist_qq(mtcars_lm, resid_type = "both")
}
\seealso{
\code{\link{resid_fitted_plot}} for residual vs. fitted values
  plots.
}
